docs: rename file
